+++ layers/Schema/packages/users/src/ModuleProcessors/FilterInnerModuleProcessor.php
@@ -12,12 +12,16 @@ class FilterInnerModuleProcessor extends AbstractModuleProcessor
 {
     public const MODULE_FILTERINNER_USERS = 'filterinner-users';
     public const MODULE_FILTERINNER_USERCOUNT = 'filterinner-usercount';
+    public const MODULE_FILTERINNER_ADMINUSERS = 'filterinner-adminusers';
+    public const MODULE_FILTERINNER_ADMINUSERCOUNT = 'filterinner-adminusercount';
 
     public function getModulesToProcess(): array
     {
         return array(
             [self::class, self::MODULE_FILTERINNER_USERS],
             [self::class, self::MODULE_FILTERINNER_USERCOUNT],
+            [self::class, self::MODULE_FILTERINNER_ADMINUSERS],
+            [self::class, self::MODULE_FILTERINNER_ADMINUSERCOUNT],
         );
     }
 
@@ -25,27 +29,35 @@ class FilterInnerModuleProcessor extends AbstractModuleProcessor
     {
         $ret = parent::getSubmodules($module);
 
-        $inputmodules = [
-            self::MODULE_FILTERINNER_USERS => [
+        $inputmodules = match ($module[1]) {
+            self::MODULE_FILTERINNER_USERS,
+            self::MODULE_FILTERINNER_ADMINUSERS => [
                 [FilterInputModuleProcessor::class, FilterInputModuleProcessor::MODULE_FILTERINPUT_NAME],
-                [FilterInputModuleProcessor::class, FilterInputModuleProcessor::MODULE_FILTERINPUT_EMAILS],
                 [CommonFilterInputModuleProcessor::class, CommonFilterInputModuleProcessor::MODULE_FILTERINPUT_ORDER],
                 [CommonFilterInputModuleProcessor::class, CommonFilterInputModuleProcessor::MODULE_FILTERINPUT_LIMIT],
                 [CommonFilterInputModuleProcessor::class, CommonFilterInputModuleProcessor::MODULE_FILTERINPUT_OFFSET],
                 [CommonFilterInputModuleProcessor::class, CommonFilterInputModuleProcessor::MODULE_FILTERINPUT_IDS],
                 [CommonFilterInputModuleProcessor::class, CommonFilterInputModuleProcessor::MODULE_FILTERINPUT_ID],
             ],
-            self::MODULE_FILTERINNER_USERCOUNT => [
+            self::MODULE_FILTERINNER_USERCOUNT,
+            self::MODULE_FILTERINNER_ADMINUSERCOUNT => [
                 [FilterInputModuleProcessor::class, FilterInputModuleProcessor::MODULE_FILTERINPUT_NAME],
-                [FilterInputModuleProcessor::class, FilterInputModuleProcessor::MODULE_FILTERINPUT_EMAILS],
                 [CommonFilterInputModuleProcessor::class, CommonFilterInputModuleProcessor::MODULE_FILTERINPUT_IDS],
                 [CommonFilterInputModuleProcessor::class, CommonFilterInputModuleProcessor::MODULE_FILTERINPUT_ID],
             ],
-        ];
+            default => [],
+        };
+        // The "email" is a restricted arg
+        if (in_array($module[1], [
+            self::MODULE_FILTERINNER_ADMINUSERS,
+            self::MODULE_FILTERINNER_ADMINUSERCOUNT,
+        ])) {
+            $inputmodules[] = [FilterInputModuleProcessor::class, FilterInputModuleProcessor::MODULE_FILTERINPUT_EMAILS];
+        }
         if (
             $modules = $this->hooksAPI->applyFilters(
                 'Users:FilterInnerModuleProcessor:inputmodules',
-                $inputmodules[$module[1]],
+                $inputmodules,
                 $module
             )
         ) {